# api/openapi/loan-context.yaml
openapi: 3.0.3
info:
  title: Loan Context Service API
  version: 1.0.0
  description: |-
    Loan origination and servicing APIs of svc-ln-loan-lifecycle (first-party and internal callers).

    Access tokens are Keycloak JWTs sent as `Authorization: Bearer`; `aud` must contain
    `svc-ln-loan-lifecycle`. Sender-constrained (DPoP) tokens are not required on these
    `/api/v1` paths: the platform contract requires DPoP only on TPP-facing endpoints, and no TPP
    calls this API. A customer (realm role customer) acts only on loans whose customerId equals the
    `customer_id` claim of the token; staff (banker, loan_officer, admin) and services (service) act
    on any loan.

    Money: every amount carries an ISO 4217 currency; there is no default currency.
  license:
    name: Proprietary
    url: https://openfinance.example.com/license
servers:
  - url: https://api.sandbox.openfinance.ae
    description: Sandbox

tags:
  - name: LoanManagement

paths:
  /api/v1/loans:
    post:
      tags: [LoanManagement]
      summary: Create loan application
      operationId: createLoanApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: INSUFFICIENT_CREDIT, CURRENCY_MISMATCH (the customer's credit is held in another currency), CUSTOMER_NOT_FOUND or CREDIT_MOVEMENT_REJECTED (the customer service rejected the credit movement as invalid; retrying will not help)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: CUSTOMER_SERVICE_UNAVAILABLE - the customer service could not be asked; nothing was changed, retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/loans/{loanId}:
    get:
      tags: [LoanManagement]
      summary: Get loan by id
      operationId: getLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/LoanId'
      responses:
        '200':
          description: Loan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/loans/{loanId}/approve:
    post:
      tags: [LoanManagement]
      summary: Approve loan
      operationId: approveLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/LoanId'
      responses:
        '200':
          description: Loan approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: INVALID_LOAN_STATE (the loan is not in a state that allows this) or CONCURRENT_UPDATE (retry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/loans/{loanId}/reject:
    post:
      tags: [LoanManagement]
      summary: Reject loan
      operationId: rejectLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/LoanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Loan rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: INVALID_LOAN_STATE (the loan is not in a state that allows this) or CONCURRENT_UPDATE (retry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: INVALID_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/loans/{loanId}/disburse:
    post:
      tags: [LoanManagement]
      summary: Disburse loan
      operationId: disburseLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/LoanId'
      responses:
        '200':
          description: Loan disbursed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: INVALID_LOAN_STATE (the loan is not in a state that allows this), CONCURRENT_UPDATE (retry) or CREDIT_RESERVATION_HELD_FOR_OPERATOR (the customer service refused to release an earlier reservation of this loan with RELEASE_EXCEEDS_RESERVATION; the loan stays APPROVED and nothing is reserved until an operator resolves it, so do not retry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                creditReservationHeldForOperator:
                  summary: The loan's credit reservation is held for an operator
                  value:
                    code: CREDIT_RESERVATION_HELD_FOR_OPERATOR
                    message: The loan's credit reservation is held for an operator; the request cannot proceed until it is resolved
                    interactionId: 6f1c2a9e-3b4d-4e8f-9a7b-1c2d3e4f5a6b
                    timestamp: '2026-10-10T09:00:00Z'
        '422':
          description: INSUFFICIENT_CREDIT (the customer service refused the reservation; the loan stays APPROVED), CURRENCY_MISMATCH, CUSTOMER_NOT_FOUND or CREDIT_MOVEMENT_REJECTED (the customer service rejected the reservation as invalid; the loan stays APPROVED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: CUSTOMER_SERVICE_UNAVAILABLE - the credit could not be reserved; the loan stays APPROVED, retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/loans/{loanId}/payments:
    post:
      tags: [LoanManagement]
      summary: Make loan payment
      operationId: makePayment
      description: |-
        Applies a repayment to the schedule in installment order, interest first, and records how it
        was allocated. When the schedule is fully paid the loan becomes FULLY_PAID and the customer's
        reserved credit is released after the repayment is stored.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/LoanId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanPaymentRequest'
      responses:
        '200':
          description: Payment applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: INVALID_REQUEST (missing or non-positive amount, currency not ISO 4217 or not the loan currency, amount above what is due, x-idempotency-key longer than 128)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: INVALID_LOAN_STATE, CONCURRENT_UPDATE or DUPLICATE_REQUEST (the same key is being processed; retry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: IDEMPOTENCY_KEY_REUSED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/loans/{loanId}/cancel:
    post:
      tags: [LoanManagement]
      summary: Cancel loan
      operationId: cancelLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/LoanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Loan cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: INVALID_LOAN_STATE (the loan is not in a state that allows this) or CONCURRENT_UPDATE (retry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: INVALID_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
    DPoP:
      name: DPoP
      in: header
      required: false
      description: Not required on these first-party and internal paths (Bearer tokens); accepted and ignored if sent.
      schema:
        type: string
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      required: false
      description: |-
        Optional. A repeat of the same request (same caller, key, loan and amount) is applied once and
        returns the loan; the same key with another loan or amount is 422 IDEMPOTENCY_KEY_REUSED.
        Keys are scoped to the caller (the customer_id claim for customers, the token subject otherwise).
      schema:
        type: string
        maxLength: 128
    InteractionId:
      name: X-FAPI-Interaction-ID
      in: header
      required: true
      schema:
        type: string
    LoanId:
      name: loanId
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateLoanRequest:
      type: object
      required: [customerId, principalAmount, currency, annualInterestRate, termInMonths]
      properties:
        customerId:
          type: string
          description: For a customer caller, must equal the customer_id claim of the token (else 403).
        principalAmount:
          type: number
          format: double
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217. Required; there is no default currency.
          example: AED
        annualInterestRate:
          type: number
          format: double
          description: Nominal annual rate in percent (6.5 = 6.5%), compounded monthly on the declining balance.
        termInMonths:
          type: integer
          minimum: 1

    LoanPaymentRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217; must be the loan's currency.
          example: AED

    ReasonRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500

    LoanResponse:
      type: object
      required: [loanId, customerId, principalAmount, annualInterestRate, rateBasis, currency, termInMonths, status, applicationDate, outstandingBalance, monthlyPayment, createdAt, lastModifiedAt]
      properties:
        loanId:
          type: string
        customerId:
          type: string
        principalAmount:
          type: number
          format: double
        annualInterestRate:
          type: number
          format: double
          description: |-
            Percent per year (6.5 = 6.5%). NOMINAL_ANNUAL: the nominal rate. FLAT_TOTAL: the flat total
            rate annualised for display, flatTotalRate * 100 * 12 / termInMonths.
        rateBasis:
          type: string
          enum: [NOMINAL_ANNUAL, FLAT_TOTAL]
          description: NOMINAL_ANNUAL for loans created by this service; FLAT_TOTAL for loans migrated from the monolith.
        flatTotalRate:
          type: number
          format: double
          nullable: true
          description: |-
            FLAT_TOTAL loans only: the monolith's interest rate as it stored it, a fraction of the principal
            charged once over the whole term (0.200 = 20%). Null for NOMINAL_ANNUAL loans.
        currency:
          type: string
          description: ISO 4217 currency of every amount of the loan.
        termInMonths:
          type: integer
        status:
          type: string
          enum: [CREATED, PENDING_APPROVAL, APPROVED, ACTIVE, DISBURSED, FULLY_PAID, DEFAULTED, REJECTED, CANCELLED, RESTRUCTURED, WRITTEN_OFF]
        applicationDate:
          type: string
          format: date
        approvalDate:
          type: string
          format: date
          nullable: true
        disbursementDate:
          type: string
          format: date
          nullable: true
        maturityDate:
          type: string
          format: date
          nullable: true
        outstandingBalance:
          type: number
          format: double
          description: Everything still due on the schedule, unpaid principal plus scheduled interest.
        monthlyPayment:
          type: number
          format: double
          description: The first installment of the schedule (for migrated loans, the monolith's installment).
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [code, message, interactionId, timestamp]
      properties:
        code:
          type: string
          description: |-
            Stable machine-readable code: INVALID_REQUEST, FORBIDDEN, LOAN_NOT_FOUND, INVALID_LOAN_STATE,
            CONCURRENT_UPDATE, DUPLICATE_REQUEST, INSUFFICIENT_CREDIT, CURRENCY_MISMATCH, CUSTOMER_NOT_FOUND,
            CREDIT_MOVEMENT_REJECTED, CREDIT_RESERVATION_HELD_FOR_OPERATOR, IDEMPOTENCY_KEY_REUSED,
            CUSTOMER_SERVICE_UNAVAILABLE, INTERNAL_ERROR.
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time